refactor(sync): use int32_t for error and sync dbIds
Replace int64_t with int32_t for error and sync dbIds across
ConflictingFilesCorrector, SyncPal, AbstractErrorResolveConflictsJob,
ErrorResolveConflictsJob, and ErrorResolveConflictsQuickJob. Update
method signatures, member variables, and local variables accordingly.
Improve error processing logic to avoid duplicate handling. Enhance
AppServer error handling for getConflictList. Adjust notification and
callback signatures to match new types. This change improves
consistency and memory usage for dbId management.

+++ src/libsyncengine/syncpal/conflictingfilescorrector.h
@@ -34,7 +34,7 @@ class ConflictingFilesCorrector : public AbstractPropagatorJob {
         ExitInfo runJob() override;
 
         uint64_t nbErrors() const { return _nbErrors; }
-        std::vector<int64_t> removedErrorsDbIds() const { return _removedErrorsDbIds; }
+        std::vector<int32_t> removedErrorsDbIds() const { return _removedErrorsDbIds; }
 
 
     private:
@@ -46,7 +46,7 @@ class ConflictingFilesCorrector : public AbstractPropagatorJob {
         std::shared_ptr<SyncPal> _syncPal = nullptr;
         const std::vector<Error> _keepLocalErrors;
         const std::vector<Error> _keepRemoteErrors;
-        std::vector<int64_t> _removedErrorsDbIds;
+        std::vector<int32_t> _removedErrorsDbIds;
         uint64_t _nbErrors = 0;
 };
